1. Modularize probability package 2. Give 50 50 probability for crossover and mutation
